Fix last details. No more duplicates.
